starting methods
Adding methods to our class
In the previous section you were introduced to the concept of an
interface - a way of interacting an object in a predefined way. The
example used was that of a television remote control, the interface it
provides is a set of buttons that you can press to change channels and
access other functionality. Behind the scenes much more is happening
but you interact with the television through this interface. The same
is true with any classes you define, you will write methods that you
and others can use to interact with objects of that type.

We are going to add two methods to our class:

Needs
Report