peta will not be happy
our grow and auto grow function work. so far not many errors.

Since my classmates have added a death feature I’ve tried adding it too
(how could everypony be so cruel)
(well it’s more of a send to a better farm in the country side
feature…)

Anyways - I’ve noticed that the Auto grow as is doesn’t have a
validator for days - so I better put that inside the method.